Replace kernel::sleep by kernel::time
