+++ src/kernel/idt.rs
@@ -12,7 +12,7 @@ lazy_static! {
         unsafe {
             idt.double_fault.set_handler_fn(double_fault_handler).set_stack_index(kernel::gdt::DOUBLE_FAULT_IST_INDEX);
         }
-        idt[IRQ0 as usize].set_handler_fn(timer_interrupt_handler);
+        idt[IRQ0 as usize].set_handler_fn(kernel::time::interrupt_handler);
         idt[IRQ1 as usize].set_handler_fn(kernel::keyboard::interrupt_handler);
         idt
     };
@@ -28,12 +28,4 @@ extern "x86-interrupt" fn breakpoint_handler(stack_frame: &mut InterruptStackFra
 
 extern "x86-interrupt" fn double_fault_handler(stack_frame: &mut InterruptStackFrame, _error_code: u64) -> ! {
     panic!("EXCEPTION: DOUBLE FAULT\n{:#?}", stack_frame);
-}
-
-extern "x86-interrupt" fn timer_interrupt_handler(_stack_frame: &mut InterruptStackFrame) {
-    kernel::clock::tick();
-
-    unsafe {
-        kernel::pic::PICS.lock().notify_end_of_interrupt(IRQ0);
-    }
 }
